Add gameplay tag to the gameplay event of dealing damage
Adding animation notify state, used to signal when melee starts and when melee ends. In this animation state we let the character know when the collider of the weapon can be enabled and when not.

When overlap of weapon happens on an enemy, it will send to the player character a ghameplay event, with the deal damage tag, and the ifnormation needed of the victim to do damage in the near future

Adding attribute set to the character base class